# Platform/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Balaagh AI trained-model analysis and persistent report API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  database?: string;
  modelsLoaded?: boolean;
  llmEnabled?: boolean;
}

export type IncidentClass = typeof IncidentClass[keyof typeof IncidentClass];


export const IncidentClass = {
  'Fire_/_Explosion': 'Fire / Explosion',
  'Flood_/_Severe_Weather': 'Flood / Severe Weather',
  'Infrastructure_/_Utilities': 'Infrastructure / Utilities',
  'Road_/_Transportation': 'Road / Transportation',
  'People_at_Risk_/_Medical': 'People at Risk / Medical',
  Other: 'Other',
} as const;

export type Priority = typeof Priority[keyof typeof Priority];


export const Priority = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
  Critical: 'Critical',
} as const;

export interface Report {
  id: number;
  originalText: string;
  incidentClass: IncidentClass;
  priority: Priority;
  /** @maxLength 5000 */
  location: string;
  peopleAtRisk: boolean;
  /** @maxLength 500 */
  requiredSupport: string;
  /** @maxLength 5000 */
  summary: string;
  createdAt: string;
  analysisTime: string;
  /** UTC ISO-8601 last-update timestamp */
  updatedAt: string;
}

export interface ReportInput {
  /**
     * @minLength 5
     * @maxLength 5000
     */
  originalText: string;
  incidentClass: IncidentClass;
  priority: Priority;
  /** @maxLength 5000 */
  location: string;
  peopleAtRisk: boolean;
  /** @maxLength 500 */
  requiredSupport: string;
  /** @maxLength 5000 */
  summary: string;
  /**
     * Optional signed model metadata receipt; preserve unchanged between analysis and save.
     * @maxLength 200000
     * @nullable
     */
  analysisToken?: string | null;
}

export interface ReportUpdate {
  incidentClass?: IncidentClass;
  priority?: Priority;
  /** @maxLength 5000 */
  location?: string;
  peopleAtRisk?: boolean;
  /** @maxLength 500 */
  requiredSupport?: string;
  /** @maxLength 5000 */
  summary?: string;
}

export interface AnalysisInput {
  /**
     * @minLength 5
     * @maxLength 5000
     */
  text: string;
}

export interface AnalysisResult {
  incidentClass: IncidentClass;
  priority: Priority;
  /** @maxLength 5000 */
  location: string;
  peopleAtRisk: boolean;
  /** @maxLength 500 */
  requiredSupport: string;
  /** @maxLength 5000 */
  summary: string;
  /**
     * Optional signed model metadata receipt; preserve unchanged between analysis and save.
     * @maxLength 200000
     * @nullable
     */
  analysisToken?: string | null;
}

export interface CountItem {
  label: string;
  count: number;
}

export interface DashboardSummary {
  totalReports: number;
  criticalReports: number;
  highPriority: number;
  reportsToday: number;
  recentReports: Report[];
  byIncidentClass: CountItem[];
  byPriority: CountItem[];
}

export interface LocationSummary {
  location: string;
  reports: number;
  criticalReports: number;
  peopleAtRisk: number;
}

export interface LocationsSummary {
  locations: LocationSummary[];
}

export interface ModelEvaluation {
  accuracy: number;
  precision: number;
  recall: number;
  macroF1: number;
  confusionMatrix: number[][];
}

export interface AnalyticsSummary {
  byIncidentClass: CountItem[];
  byPriority: CountItem[];
  byLocation: CountItem[];
  bySupport: CountItem[];
  peopleAtRisk: number;
  evaluation: ModelEvaluation | null;
}

export interface ReportEdit {
  reportId: number;
  editedAt: string;
  field: string;
  before: string;
  after: string;
  originalText: string;
  incidentClass: IncidentClass;
}

export type ApiErrorError = {
  code: string;
  message: string;
  requestId: string;
};

export interface ApiError {
  error: ApiErrorError;
}

export type SearchParameter = string;

export type IncidentClassParameter = IncidentClass;

export type PriorityParameter = Priority;

export type LocationParameter = string;

export type SortParameter = typeof SortParameter[keyof typeof SortParameter];


export const SortParameter = {
  date: 'date',
  priority: 'priority',
  incidentClass: 'incidentClass',
} as const;

export type ListReportsParams = {
search?: SearchParameter;
incidentClass?: IncidentClassParameter;
priority?: PriorityParameter;
location?: LocationParameter;
sort?: SortParameter;
};