# Platform/lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Balaagh AI trained-model analysis and persistent report API
 * OpenAPI spec version: 1.0.0
 */
import type { CountItem } from './countItem';
import type { ModelEvaluation } from './modelEvaluation';

export interface AnalyticsSummary {
  byIncidentClass: CountItem[];
  byPriority: CountItem[];
  byLocation: CountItem[];
  bySupport: CountItem[];
  peopleAtRisk: number;
  evaluation: ModelEvaluation | null;
}