# Platform/lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Balaagh AI trained-model analysis and persistent report API
 * OpenAPI spec version: 1.0.0
 */
import type { ApiErrorError } from './apiErrorError';

export interface ApiError {
  error: ApiErrorError;
}