# Platform/lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Balaagh AI trained-model analysis and persistent report API
 * OpenAPI spec version: 1.0.0
 */
import type { IncidentClass } from './incidentClass';
import type { Priority } from './priority';

export interface ReportInput {
  /**
     * @minLength 5
     * @maxLength 5000
     */
  originalText: string;
  incidentClass: IncidentClass;
  priority: Priority;
  /** @maxLength 5000 */
  location: string;
  peopleAtRisk: boolean;
  /** @maxLength 500 */
  requiredSupport: string;
  /** @maxLength 5000 */
  summary: string;
  /**
     * Optional signed model metadata receipt; preserve unchanged between analysis and save.
     * @maxLength 200000
     * @nullable
     */
  analysisToken?: string | null;
}